+++ apps/demo/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @apps/demo
 
+## 5.1.2
+
+### Patch Changes
+
+- Added Solana Auth Provider types and schemas
+
+- Updated dependencies []:
+  - @web3modal/wagmi@5.1.2
+
 ## 5.1.1
 
 ### Patch Changes